third_party/blink/web_tests/inspector-protocol/css/css-set-rule-selector-expected.txt

Tests CSS.setRuleSelector method
==== Initial style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter { font-size: 200%; } div,
body {
    position: relative;
}



Running test: testEditSimpleSelector
==== Style sheet text ====
.EDITED-SELECTOR {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter { font-size: 200%; } div,
body {
    position: relative;
}



Running test: testEditCopmlexSelector
==== Style sheet text ====
* {
    box-sizing: border-box;
}

.EDITED-SELECTOR:first-line {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

p:first-letter { font-size: 200%; } div,
body {
    position: relative;
}



Running test: testEditSequentially
==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

.EDITED { font-size: 200%; } div,
body {
    position: relative;
}


==== Style sheet text ====
* {
    box-sizing: border-box;
}

html, body, div, a {
    /* resetting some properties */
    padding: 0;
    margin: 0;
    border: 0;
    display: flex; /* flex FTW! */
}

.EDITED { font-size: 200%; } .EDITED-2 {
    position: relative;
}



Running test: testInvalidSelectorText
Expected protocol error: SyntaxError Selector or media text is not valid.

Running test: testEmptySelector
Expected protocol error: SyntaxError Selector or media text is not valid.

Running test: testSelectorTextWithUnclosedComment
Expected protocol error: SyntaxError Selector or media text is not valid.

Running test: testSelectorTextWithBrace
Expected protocol error: SyntaxError Selector or media text is not valid.

Running test: testSelectorTextWithBraces
Expected protocol error: SyntaxError Selector or media text is not valid.

Running test: testSelectorTextWithExtraRule
Expected protocol error: SyntaxError Selector or media text is not valid.

Running test: testEditSubSelector
Expected protocol error: NotFoundError Source range didn't match existing source range

Running test: testInvalidParameters
Expected protocol error: Invalid parameters (Failed to deserialize params.range.startLine - BINDINGS: int32 value expected <somewhere>)

Running test: testNegativeRangeParameters
Expected protocol error: range.startLine must be a non-negative integer

Running test: testStartLineOutOfBounds
Expected protocol error: Specified range is out of bounds

Running test: testEndLineOutOfBounds
Expected protocol error: Specified range is out of bounds

Running test: testStartColumnBeyondLastLineCharacter
Expected protocol error: Specified range is out of bounds

Running test: testEndColumnBeyondLastLineCharacter
Expected protocol error: Specified range is out of bounds

Running test: testInsertBeyondLastCharacterOfLastLine
Expected protocol error: Specified range is out of bounds

Running test: testReversedRange
Expected protocol error: Range start must not succeed its end
